+++ src/app/app-routing.module.ts
@@ -1,5 +1,5 @@
 import { NgModule } from "@angular/core";
-import { RouterModule, Routes } from "@angular/router";
+import { RouterModule, Routes, ExtraOptions } from "@angular/router";
 import { ChatComponent } from "./chat/chat.component";
 import { PortfolioComponent } from "./portfolio/portfolio.component";
 
@@ -15,8 +15,12 @@ const routes: Routes = [
   }
 ];
 
+const routerOptions: ExtraOptions = {
+  anchorScrolling: "enabled"
+};
+
 @NgModule({
-  imports: [RouterModule.forRoot(routes)],
+  imports: [RouterModule.forRoot(routes, routerOptions)],
   exports: [RouterModule]
 })
 export class AppRoutingModule {}